pyproject.toml
[tool.poetry]
name = "bqunit"
version = "0.1.0"
description = ""
authors = ["to-lz1 <m.toriyama000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-bigquery = "^1.24.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
